blocksクラスの削除
