+++ app/src/main/java/com/example/tetris/MainActivity.java
@@ -1,13 +1,5 @@
 package com.example.tetris;
 
-import static com.example.tetris.blocks.iBlock;
-import static com.example.tetris.blocks.jBlock;
-import static com.example.tetris.blocks.lBlock;
-import static com.example.tetris.blocks.oBlock;
-import static com.example.tetris.blocks.sBlock;
-import static com.example.tetris.blocks.tBlock;
-import static com.example.tetris.blocks.zBlock;
-
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Context;
@@ -45,19 +37,25 @@ public class MainActivity extends AppCompatActivity {
     int offsety=0;
 
     int[][] field=new int[15][10];
-
+    public static final int tBlock = 1;
+    public static final int sBlock = 2;
+    public static final int iBlock = 3;
+    public static final int oBlock = 4;
+    public static final int lBlock = 5;
+    public static final int jBlock = 6;
+    public static final int zBlock = 7;
 
     int motion=Stational;
     public static final int Stational=0;
     public static final int Left=1;
     public static final int Right=2;
 
-
+    int num=0;
 
     public static final int blockLenght=4;
+    int[][] nowBlock=new int[blockLenght][blockLenght];
 
-
-    public static boolean moveflag;
+    boolean moveflag;
 
 
     @Override
@@ -74,7 +72,7 @@ public class MainActivity extends AppCompatActivity {
         protected void onDraw(Canvas ca) {
             super.onDraw(ca);
             blockDraw bd=new blockDraw();
-            com.example.tetris.blocks bs=new com.example.tetris.blocks();
+            blocks bs=new blocks();
             Reset re=new Reset();
             setTimer st=new setTimer();
             movement mm=new movement();
@@ -149,7 +147,7 @@ public class MainActivity extends AppCompatActivity {
                 //ブロック出現処置
                 for(int i=0;i<blockLenght;i++) {
                     for (int j = 0; j < blockLenght; j++) {
-                        if (com.example.tetris.blocks.nowBlock[i][j] == 1) {
+                        if (nowBlock[i][j] == 1) {
                             drawMoveBlock(+offsetx + j, offsety + i, ca);
                         }
                     }
@@ -167,7 +165,7 @@ public class MainActivity extends AppCompatActivity {
                 int px=x*blocksize;
                 int py=y*blocksize;
 
-                switch (com.example.tetris.blocks.num){
+                switch (num){
                     case tBlock:
                         paint.setColor(Color.BLUE);
                         ca.drawRect(px+blocksize,py+blocksize,px,py,paint);
@@ -248,7 +246,78 @@ public class MainActivity extends AppCompatActivity {
 
     class blocks{
 
+        public void blocks(){
 
+            //ブロックランダム抽選
+            num=(int)(Math.random()*7)+1;
+
+            switch (num){
+                case tBlock:
+                    nowBlock=new int[][]{
+                            {0, 0, 0, 0},
+                            {0, 0, 1, 0},
+                            {0, 1, 1, 1},
+                            {0, 0, 0, 0}
+                    };
+                    break;
+
+                case sBlock:
+                    nowBlock=new int[][]{
+                            {0, 0, 0, 0},
+                            {0, 1, 1, 0},
+                            {1, 1, 0, 0},
+                            {0, 0, 0, 0}
+                    };
+                    break;
+
+                case iBlock:
+                    nowBlock=new int[][]{
+                            {0, 1, 0, 0},
+                            {0, 1, 0, 0},
+                            {0, 1, 0, 0},
+                            {0, 1, 0, 0}
+                    };
+                    break;
+
+                case oBlock:
+                    nowBlock=new int[][]{
+                            {0, 1, 1, 0},
+                            {0, 1, 1, 0},
+                            {0, 0, 0, 0},
+                            {0, 0, 0, 0}
+                    };
+                    break;
+
+                case lBlock:
+                    nowBlock=new int[][]{
+                            {0, 0, 0, 0},
+                            {0, 1, 1, 1},
+                            {0, 1, 0, 0},
+                            {0, 0, 0, 0}
+                    };
+                    break;
+
+                case jBlock:
+                    nowBlock=new int[][]{
+                            {0, 1, 0, 0},
+                            {0, 1, 1, 1},
+                            {0, 0, 0, 0},
+                            {0, 0, 0, 0}
+                    };
+                    break;
+                case zBlock:
+                    nowBlock=new int[][]{
+                            {0, 0, 0, 0},
+                            {0, 1, 1, 0},
+                            {0, 0, 1, 1},
+                            {0, 0, 0, 0}
+                    };
+                    break;
+                default:
+                    break;
+            };
+            moveflag=true;
+        }
     }
 
     class Reset{
